# Server/Controllers/FinancialStatementApi.cs
/*
 * StockQuotes
 *
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 0.1
 * 
 * Generated by: https://github.com/swagger-api/swagger-codegen.git
 */
using Microsoft.AspNetCore.Mvc;
using Newtonsoft.Json;
using Server.Attributes;
using Swashbuckle.AspNetCore.Annotations;
using System;
using System.ComponentModel.DataAnnotations;

namespace Server.Controllers {
	/// <summary>
	/// 
	/// </summary>
	[ApiController]
	public class FinancialStatementApiController : ControllerBase {
		/// <summary>
		/// 
		/// </summary>
		/// <param name="id"></param>
		/// <param name="year">Default value is current year</param>
		/// <param name="quarter">Default value is current quarter</param>
		/// <response code="200">Success</response>
		[HttpGet]
		[Route("/truemogician/StockQuote/0.1/api/statement/cash-flow")]
		[ValidateModelState]
		[SwaggerOperation("GetCashFlow")]
		public virtual IActionResult GetCashFlow([FromQuery][Required()][RegularExpression("/\\d{6}\\.[a-zA-Z]{2}/")] string id, [FromQuery] int? year, [FromQuery] int? quarter) {
			//TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
			// return StatusCode(200);

			throw new NotImplementedException();
		}

		/// <summary>
		/// 
		/// </summary>
		/// <param name="id"></param>
		/// <param name="year">Default value is current year</param>
		/// <param name="quarter">Default value is current quarter</param>
		/// <response code="200">Success</response>
		[HttpGet]
		[Route("/truemogician/StockQuote/0.1/api/statement/dupond")]
		[ValidateModelState]
		[SwaggerOperation("GetDuPond")]
		public virtual IActionResult GetDuPond([FromQuery][Required()][RegularExpression("/\\d{6}\\.[a-zA-Z]{2}/")] string id, [FromQuery] int? year, [FromQuery] int? quarter) {
			//TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
			// return StatusCode(200);

			throw new NotImplementedException();
		}

		/// <summary>
		/// 
		/// </summary>
		/// <param name="id"></param>
		/// <param name="year">Default value is current year</param>
		/// <param name="quarter">Default value is current quarter</param>
		/// <response code="200">Success</response>
		[HttpGet]
		[Route("/truemogician/StockQuote/0.1/api/statement/growth")]
		[ValidateModelState]
		[SwaggerOperation("GetGrowthAbility")]
		public virtual IActionResult GetGrowthAbility([FromQuery][Required()][RegularExpression("/\\d{6}\\.[a-zA-Z]{2}/")] string id, [FromQuery] int? year, [FromQuery] int? quarter) {
			//TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
			// return StatusCode(200);

			throw new NotImplementedException();
		}

		/// <summary>
		/// 
		/// </summary>
		/// <param name="id"></param>
		/// <param name="year">Default value is current year</param>
		/// <param name="quarter">Default value is current quarter</param>
		/// <response code="200">Success</response>
		[HttpGet]
		[Route("/truemogician/StockQuote/0.1/api/statement/operation")]
		[ValidateModelState]
		[SwaggerOperation("GetOperationCapability")]
		public virtual IActionResult GetOperationCapability([FromQuery][Required()][RegularExpression("/\\d{6}\\.[a-zA-Z]{2}/")] string id, [FromQuery] int? year, [FromQuery] int? quarter) {
			//TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
			// return StatusCode(200);

			throw new NotImplementedException();
		}

		/// <summary>
		/// 
		/// </summary>
		/// <param name="id"></param>
		/// <param name="beginDate"></param>
		/// <param name="endDate"></param>
		/// <response code="200">Success</response>
		[HttpGet]
		[Route("/truemogician/StockQuote/0.1/api/statement/forcast")]
		[ValidateModelState]
		[SwaggerOperation("GetPerformanceForcast")]
		public virtual IActionResult GetPerformanceForcast([FromQuery][Required()][RegularExpression("/\\d{6}\\.[a-zA-Z]{2}/")] string id, [FromQuery] DateTime? beginDate, [FromQuery] DateTime? endDate) {
			//TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
			// return StatusCode(200);

			throw new NotImplementedException();
		}

		/// <summary>
		/// 
		/// </summary>
		/// <param name="id"></param>
		/// <param name="beginDate"></param>
		/// <param name="endDate"></param>
		/// <response code="200">Success</response>
		[HttpGet]
		[Route("/truemogician/StockQuote/0.1/api/statement/report")]
		[ValidateModelState]
		[SwaggerOperation("GetPerformanceReport")]
		public virtual IActionResult GetPerformanceReport([FromQuery][Required()][RegularExpression("/\\d{6}\\.[a-zA-Z]{2}/")] string id, [FromQuery] DateTime? beginDate, [FromQuery] DateTime? endDate) {
			//TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
			// return StatusCode(200);

			throw new NotImplementedException();
		}

		/// <summary>
		/// 
		/// </summary>
		/// <param name="id"></param>
		/// <param name="year">Default value is current year</param>
		/// <param name="quarter">Default value is current quarter</param>
		/// <response code="200">Success</response>
		[HttpGet]
		[Route("/truemogician/StockQuote/0.1/api/statement/profit")]
		[ValidateModelState]
		[SwaggerOperation("GetProfitability")]
		public virtual IActionResult GetProfitability([FromQuery][Required()][RegularExpression("/\\d{6}\\.[a-zA-Z]{2}/")] string id, [FromQuery] int? year, [FromQuery] int? quarter) {
			//TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
			// return StatusCode(200);

			throw new NotImplementedException();
		}

		/// <summary>
		/// 
		/// </summary>
		/// <param name="id"></param>
		/// <param name="year">Default value is current year</param>
		/// <param name="quarter">Default value is current quarter</param>
		/// <response code="200">Success</response>
		[HttpGet]
		[Route("/truemogician/StockQuote/0.1/api/statement/balance")]
		[ValidateModelState]
		[SwaggerOperation("GetSolvency")]
		public virtual IActionResult GetSolvency([FromQuery][Required()][RegularExpression("/\\d{6}\\.[a-zA-Z]{2}/")] string id, [FromQuery] int? year, [FromQuery] int? quarter) {
			//TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
			// return StatusCode(200);

			throw new NotImplementedException();
		}
	}
}