# Server/wwwroot/swagger-original.json
{
  "openapi" : "3.0.0",
  "info" : {
    "title" : "StockQuotes",
    "version" : "0.1"
  },
  "servers" : [ {
    "url" : "https://virtserver.swaggerhub.com/truemogician/StockQuote/0.1",
    "description" : "SwaggerHub API Auto Mocking"
  } ],
...
  "paths" : {
...
    "/api/statement/profit" : {
      "get" : {
        "tags" : [ "Financial Statement" ],
        "operationId" : "GetProfitability",
        "parameters" : [ {
          "name" : "id",
          "in" : "query",
          "required" : true,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "pattern" : "\\d{6}\\.[a-zA-Z]{2}",
            "type" : "string"
          }
        }, {
          "name" : "year",
          "in" : "query",
          "description" : "Default value is current year",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "name" : "quarter",
          "in" : "query",
          "description" : "Default value is current quarter",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "integer",
            "enum" : [ 1, 2, 3, 4 ]
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Success"
          }
        }
      }
    },
    "/api/statement/operation" : {
      "get" : {
        "tags" : [ "Financial Statement" ],
        "operationId" : "GetOperationCapability",
        "parameters" : [ {
          "name" : "id",
          "in" : "query",
          "required" : true,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "pattern" : "\\d{6}\\.[a-zA-Z]{2}",
            "type" : "string"
          }
        }, {
          "name" : "year",
          "in" : "query",
          "description" : "Default value is current year",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "name" : "quarter",
          "in" : "query",
          "description" : "Default value is current quarter",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "integer",
            "enum" : [ 1, 2, 3, 4 ]
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Success"
          }
        }
      }
    },
    "/api/statement/growth" : {
      "get" : {
        "tags" : [ "Financial Statement" ],
        "operationId" : "GetGrowthAbility",
        "parameters" : [ {
          "name" : "id",
          "in" : "query",
          "required" : true,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "pattern" : "\\d{6}\\.[a-zA-Z]{2}",
            "type" : "string"
          }
        }, {
          "name" : "year",
          "in" : "query",
          "description" : "Default value is current year",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "name" : "quarter",
          "in" : "query",
          "description" : "Default value is current quarter",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "integer",
            "enum" : [ 1, 2, 3, 4 ]
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Success"
          }
        }
      }
    },
    "/api/statement/balance" : {
      "get" : {
        "tags" : [ "Financial Statement" ],
        "operationId" : "GetSolvency",
        "parameters" : [ {
          "name" : "id",
          "in" : "query",
          "required" : true,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "pattern" : "\\d{6}\\.[a-zA-Z]{2}",
            "type" : "string"
          }
        }, {
          "name" : "year",
          "in" : "query",
          "description" : "Default value is current year",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "name" : "quarter",
          "in" : "query",
          "description" : "Default value is current quarter",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "integer",
            "enum" : [ 1, 2, 3, 4 ]
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Success"
          }
        }
      }
    },
    "/api/statement/cash-flow" : {
      "get" : {
        "tags" : [ "Financial Statement" ],
        "operationId" : "GetCashFlow",
        "parameters" : [ {
          "name" : "id",
          "in" : "query",
          "required" : true,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "pattern" : "\\d{6}\\.[a-zA-Z]{2}",
            "type" : "string"
          }
        }, {
          "name" : "year",
          "in" : "query",
          "description" : "Default value is current year",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "name" : "quarter",
          "in" : "query",
          "description" : "Default value is current quarter",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "integer",
            "enum" : [ 1, 2, 3, 4 ]
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Success"
          }
        }
      }
    },
    "/api/statement/dupond" : {
      "get" : {
        "tags" : [ "Financial Statement" ],
        "operationId" : "GetDuPond",
        "parameters" : [ {
          "name" : "id",
          "in" : "query",
          "required" : true,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "pattern" : "\\d{6}\\.[a-zA-Z]{2}",
            "type" : "string"
          }
        }, {
          "name" : "year",
          "in" : "query",
          "description" : "Default value is current year",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "name" : "quarter",
          "in" : "query",
          "description" : "Default value is current quarter",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "integer",
            "enum" : [ 1, 2, 3, 4 ]
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Success"
          }
        }
      }
    },
    "/api/statement/report" : {
      "get" : {
        "tags" : [ "Financial Statement" ],
        "operationId" : "GetPerformanceReport",
        "parameters" : [ {
          "name" : "id",
          "in" : "query",
          "required" : true,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "pattern" : "\\d{6}\\.[a-zA-Z]{2}",
            "type" : "string"
          }
        }, {
          "name" : "begin-date",
          "in" : "query",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string",
            "format" : "date-time"
          }
        }, {
          "name" : "end-date",
          "in" : "query",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string",
            "format" : "date-time"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Success"
          }
        }
      }
    },
    "/api/statement/forcast" : {
      "get" : {
        "tags" : [ "Financial Statement" ],
        "operationId" : "GetPerformanceForcast",
        "parameters" : [ {
          "name" : "id",
          "in" : "query",
          "required" : true,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "pattern" : "\\d{6}\\.[a-zA-Z]{2}",
            "type" : "string"
          }
        }, {
          "name" : "begin-date",
          "in" : "query",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string",
            "format" : "date-time"
          }
        }, {
          "name" : "end-date",
          "in" : "query",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string",
            "format" : "date-time"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Success"
          }
        }
      }
    }
  },
  "components" : {
...
    "securitySchemes" : {
      "token" : {
        "type" : "apiKey",
        "name" : "token",
        "in" : "header"
      }
    }
  }
}